Use the finfo functions instead of apache_lookup_uri()
This function was used to lookup mimetypes when serving static files.
However, it is only available on Apache, whereas the finfo functions are
always available. The latter use content-based matching, which might be
more reliable, but it fails to recognize CSS and JS files, so some
extension-based special case is added for those.