Limit direct file serving in index.php
Previously, it would serve data/hypha.css and any file inside system
(except for system/datatypes/). However, this would also serve .php
files and perhaps other files that are not intended to be exposed in
this way. This commit reduces this feature to just the css file and
files inside system/wymeditor.