Use a nowdoc for some CSS
Previously, this used a single string and the PHP_EOL constant to insert
line endings, making the code really dense. Also, PHP_EOL is the line
ending for the platform running the server, which is not really useful
for sending content to the client.

This uses a PHP nowdoc to make the CSS more readable.