Improve password hashing
This used to apply a plain md5 hash with no salt or iteration and
non-constant-time comparison, which is generally considered unsafe. This
commit makes it use the PHP password hashing framework that was
introduced in PHP 5.5, which uses bcrypt / blowfish using a salt,
multiple iterations and constant-time comparisons, which should be a lot
more secure.

Any existing passwords, and systems using PHP < 5.5 will continue to use
the md5 scheme, but when an existing user changes his or her password,
it will be converted to the new scheme instead.