Close the PHP session as soon as possible
While the PHP session is opened, the corresponding session file is
locked. This prevents multiple requests in the same session from being
handled concurrently. By closing the session right after loading it,
it is unlocked, allowing other sessions to read it. When the session
needs to be modified, it is reloaded using session_start(), changed,
written and closed again.