Fix building of hypha.php
In 70f98ef (Move main code below function definitions), some code was
placed below the data function in hypha.php. During build, the data
function was replaced with new data, but this assumed the data function
was last in the file, since it replaced everything after it as well.

This changes the build process to only replace the contents of the
switch inside the data function, as indicated by START_OF_DATA and
END_OF_DATA markers, so other things can come after that function as
well.