Add schemaVersion attribute to hypha and page xml files
This should help with upgrades in the future. For now, this just
hardcoded at the place the XML is generated, and no checks against the
version exist, but this will be improved later (once a bit more is clear
about how modules will be split and how they will be upgraded).