Make HTMLDocument::__construct more clear
Instead of using func_num_args() to allow an optional filename to be
specified, use an explicit argument with a default value. This makes it
more clear what happens.

Also, by moving the loading of the file out of the loadHTML() call into
a separate if statement, it is a bit more clear what is happening.