Fix regex in dewikify
This regex is intended to match <a href="hypha:1234"/> links. However,
it could also match things like <a href="hypha:1234"></a><br/>. Where
the closing end of the br tag was seen as the self-closing part of the a
tag. There was already a "non-greedy" modifier on the .* in the regexes,
but in this case this was not enough. By changing it to [^/>]*, that part
of the regex will certainly terminate on the next / or >, whichever
comes first.

Eventually, this entire regex business should probably be replaced by
some more robust DOM iteration, but for now this should fix this
problem.