Lock files to prevent race conditions
Previously, files were read and written without considering any other
requests running at the same time. This could, occasionally cause some
problems:
 1. A request would read a file when another request was halfway writing
    it, leading to a single broken view.
 2. Two requests could write to the same file at the same time. The
    resulting file could be either one of the write attempts, but also a
    mix of both, depending on timing, which results in a corrupt data
    file.
 3. When a request reads a file, makes changes and then writes the
    result, it could overwrite changes from another request that were
    written while making changes.

This is fixed by using PHP's flock() function, which allows
cross-platform locking. All reads and writes to file now happen with an
appropriate lock held. Just locking during the read and write operations
would fix issue #1 and #2 above. To fix #3, it is additionally important
that an exclusive lock is held from before reading a file, until after
writing it.

To facilitate the latter, some changes have been made to the structure
of the code. When initially reading the hypha.xml file, it is not known
whether it will be written later, so it is read using a shared lock. If
it turns out changes are needed, the file is reloaded using an exclusive
lock (using Xml::lockAndReload), which is held until the changes have
been written (using Xml::saveAndUnlock).

The various hypha_setXXX functions have been changed to only make the
changes in memory, without actually writing out the changes. This moves
the responsibility of writing out changes, but also taking a lock, to
the caller. This prevents writing out hypha.xml multiple times when
making changes to multiple fields, but also ensures that these changes
are made atomatically (due to the lock). All hypha_setXXX functions now
throw an error if they are called without locking $hyphaXml first.

One open issue here is that requiring callers to call
$hyphaXml->lockAndReload() before making changes somewhat breaks the
abstraction offers by the hypha_setXXX functions. This might need some
further restructuring to fix more elegantly.

The actual locking, reading and writing of files is handled by the new
HyphaFile class, which is used by the Xml class.  For the other files
(HTML, CSS, digest, stats), HyphaFile is used directly. For each of
them, a single global HyphaFile instance is created, which can be
accessed through the new Hypha class (which should contain more of the
global state in the future).

The stats and digest APIs have been made a bit more specific, moving the
the read-modify-write cycle, as well as the needed locking into the
functions in base.php (which would not be possible with the previous
generic get and set functions, without having the caller access the
HyphaFile instance directly, which breaks the abstraction somewhat).

In the Page/TextPage classes, a reference to the Xml node in the page
list is kept. If the $hyphaXml is reloaded to take a lock, this will
still keep a reference to the old Xml node, which can break in some
cases. In TextPage::save() this is solved by replacing the node
reference with the new version, as well as updating any property values
copied out of the node. This solves the immediate problem, but is
probably not very robust - what if some other code reloads the Xml? This
might need a bit more restructuring to be properly solved.