Change display_errors based on DEBUG, not error_reporting
This means that DEBUG now controls whether messages are shown in the
browser or not, without affecting what messages are actually displayed
and/or logged (through error_reporting, which is now kept at the default
from php.ini).

Previously, when DEBUG was not set, all messages were hidden using
error_reporting, which causes them to be hidden in the browser as well
as in the webserver log, making it pretty hard to diagnose an issue.
Now, messages are always shown in the log, just disabled in the browser.