Do not pass any id to loadUser()
This changes the meaning of loadUser to load the current user from the
session, instead of the specified user. This means that index.php no
longer needs knowledge about the hyphaUser session variable, which is
now only referenced from users.php.

Additionally, by using isset, this prevents a PHP notice from appearing
when the session variable is not set.