Use a HTTP redirect for reloading after a command
Previously, when a registered command returned 'reload', only loadUser()
and loadPage() would be called again, without reloading the entire
request. By using a HTTP redirect, the browser will replace the (now
handled) POST request by a GET request, which prevents resubmitting the
POST data when the user presses refresh (and the annoying dialog that
goes along with it).

To make sure that any notifications that resulted from the command
handling are not lost, these are saved in the session so they can be
reloaded when processing the redirected request. By storing them in the
session, they can be removed once shown, to prevent the messages from
being shown again when the user refreshes. To make sure the right
messages end up in the right browser window, a unique identifier is
passed in the url, and the messages in the session are indexed by the
same identifier.