Split off language loading into new loadLanguage function
This splits reading the language out of the url from reading the
requested module or page. This prepares for a subsequent commit, where
a command needs to run before loading the requested page, but needs
access to the current language.

This generalizes language loading to apply for all requests, not just
page requests. This means you can use an url like /en/settings to switch
the current language (though the *interface* language is always taken
from the user preferences, so this language change is only visible on
the next page visit).