Let command return an array for redirecting
Previously, they returned a string that contained the redirect url
prefixed with "redirect:", which was again split by the handling code.
This works, but encoding data in a string is of course a bit fragile, so
this replaces that by returning an array where the first element is
"redirect" and the second is the url to redirect to.