Do not reset (admin) rights when reregistering
The handling of the register command always set the rights field to
'user'. This is needed on initial registration, when the rights field
contains 'invitee'. However this had the side-effect that using the
reregister feature to reset a forgotten password would remove any admin
rights for the user.

Now, the rights field is only modified when its current value is
'invitee'.