Fix reregister link in the email
In 10111b2 (Fix the reregistration process), this link was changed from
"register" to "reregister", since it appeared a separate reregister page
would be needed. However, it turned out this was not needed and the
existing register page suffices, but this link was still changed, now
pointing to a non-existing page. The displayed text did contain the
correct url.