Fix inviting multiple users
When inviting a user, a new user entry was created. hypha_addUser()
would check that no existing user with the same username or email
address would already exist. However, when inviting a user, the username
would be left empty. This means that when a second user was invited
before the first one had accepted, hypha_addUser() would refuse to add
the new user, thinking the (empty) username already existed.

Now, the username and email address are only checked for uniqueness when
they are actually given, and not when they are empty.