Fix display of invitees and exmembers in the memberlist
The memberlist contained some code to put invitees and exmembers at the
top of the list. However, this code would cause multiple invitees and
exmembers to overwrite each other, showing only the last one.
Additionally, the check for invitees was wrong, it checked for 'invite'
rights, instead of 'invitee' (but because all invitees have the same
empty username, they would still overwrite each other).

This commit fixes this by setting an explicit "priority" for each user,
and prefixing that to the username as the sort key. Additionally, the
e-mail address is added to the sort key, since invitees will have an
empty username.