Add PHP DOM Wrapper and CssSelector library
This adds some wrapper classes around the PHP DOM classes that add
additional methods for DOM querying, traversal and manipulation. These
methods mostly follow the API offered by jquery. The CssSelector library
is a dependency of DOM Wrapper, that allows using CSS-style selectors to
traverse the DOM.

By letting the Hypha Xml and Html classes derive from the DOM Wrapper
Document class, all DOM nodes will automatically be wrapped and these
methods can be used without further changes.

This does not actually make use of these new methods yet, but in the
future some Hypha helper methods can be replaced or simplified.

This imports version 0.6.3 of the DOM Wrapper library from
https://github.com/scotteh/php-dom-wrapper and version 2.8.7 from
https://github.com/symfony/css-selector/. Just the relevant
subdirectories are imported, plus the licenses (moved into a README
file). Since the libraries are originally intended to be used with
Composer and a class autoloader, an All.php is added that simply
includes all files (with a bit of trial and error for the correct
order).