Remove special case for index language
Previously, the index would use an url like /index/en to select the
index in a different language. However, since a recent refactor, any
language at the start of the url would already be parsed and saved into
$hyphaLanguage, allowing the index to make use of this.

This commit removes the special handling of languages in index urls,
which changes the url of the index to /en/index instead, which is more
consistent with other pages.