Pass all url arguments to pages
Before, only a single url component (the "view") was passed to a page
(the settings page additionally got passed a "username" component as
well). With this change, an array of components is passed instead,
allowing a page to take as much url parameters as it wants.

This removes the need for the `SettingsPage` to separately store a
`username` field, since its parent class `Page` already stores all
arguments so they can be recalled when needed. This additionally means
that the `key` argument from the register url no longer needs to be
stored inside the `username` field, which was a bit wrong anyway.