Fix settings for invitee users
Previously, they would see an "edit" button, which would not work for
lack of a username. Now, the edit button is hidden for invitees
(nothing to edit anyway). Previously, a "make admin" button would be
shown and the "remind button" was not, due to an "invite" vs "invitee"
typo, which is now fixed. Finally, the remind button would not work,
because the action name was incorrect.