Let sendMail() figure out the from address
All calls to sendMail() used the same from address, based on the hypha
settings. Instead of duplicating the construction of the from header
everywhere, this just lets sendMail() construct it.

At the same time, this adds some quoating to the From header, to prevent
spaces and quotes in the hypha title from causing problems. This quoting
might not be fool-proof in the case of non-ASCII characters, so using
mb_encode_mimeheader() could be considered in the future.