Use multipart/form-data for the main hypha form
This allows file uploads as part of this form, preparing for image
uploads in the festival module. In some existing forms, the enctype is
changed dynamically through javascript, but just using multipart always
seems easier. This will make form submissions that do not need it a bit
bigger, but that is probably not a problem in practice.