Specify a envelope-from address for sent mail
The configured system email address was already put in the From: header,
but the envelope from address was left at the default. On most systems
that ends up as username@hostname, which is not always a valid e-mail
adress, and can thus lead to email being rejected.

An extra advantage of setting the envelope address is that e-mail
bounces will also be sent to that address, instead of being sent to
whatever is the default envelope address.