festival: Allow logged in users to edit contributions
Previously, they could only add new contributions, but not edit them.
Now, an edit button is added to the list of contributions, and logged in
users can edit contributions without knowing the key. This slightly
changes the way adding contributions by logged in users works too, as
well as checkKeyArguments. It's still a bit messy, though.

As a side effect, this changes fixes a small problem where adding
contribution by a logged in user would show a notification about not
having a valid key.