festival: improve description escaping
Previously, this rendered the inner html as-is, even though it was saved
as a text node. This copies any escaping from the XML into the HTML,
which is practicallly correct, but not really the right way. This commit
uses the text() method to "undo" the XML escaping and using
htmlspecialchars to escape it into the HTML again.