Improve .htaccess rewriting
This prevents urls from being handled by index.php when they point to a valid
file or directory, which allows adding extra php scripts or files to the hypha
directory and have them served by Apache normally.

This would also cause the system and data directory from becoming available
directly, which is a problem. To prevent that, extra rules are added to let
these directories always be handled by index.php, which selectively allows some
urls but not others.