Fix the wymeditor iframe stylesheet
The actual editing in wymeditor happens in an iframe, which is intended to be
preloaded with a tiny HTML page that includes a wymeditor stylesheet. This
stylesheet emphasizes the document structure, which should encourage people to
generate clean page content.

However, the url to this iframe was broken since some time, which caused the
stylesheet to be missing and the editor look more like a normal WYSYWIG editor.
This fixes the iframe url and thus the stylesheet.