Use relative urls in more places
Previously, urls output into the main HTML were always made absolute.
Now that there is a <base> tag in the head, it's possible to make them
relative. Urls related to the editor are still absolute, since they are
also used in dialogs and the editor iframe, which does not have the base
tag (yet).