Improve generation of previous-version-diffs
When saving a page, the current version is stored in a new version node,
and the previous version node is replaced by a diff.

Previously, this diff was generated against the HTML sent by the
browser, but when this HTML is inserted into the XML, it is slightly
modified (e.g. whitespace changes), breaking the diff.

Now, the new HTML is inserted into the XML and extracted again (in the
same way that it will be extacted later), guaranteeing that the diff
still applies then.