Fix making pages public again
In d6c566a (Fixed bug that made translations save as private page), and
(accidentally) in 3351ae7 (Fix bug that showed Revert button when
selecting most recent version) the handling of the private flag was
changed, but this resulted in pages always becoming private when saving
hem. This commit reverts those fixes and fixes that same bug by letting
the translation form mimic the edit form by omitting the textPrivate
hidden field when the page is not private. This works just like the
checkbox in the edit form, which is also omitted from the request when
it is unchecked.

At the same time, one place where Page::privateFlag was referenced is
updated so it expects that variable to be a boolean instead of a string
containing 'on' or 'off'.