Preserve custom link texts and titles
Previously, during wikify the link text and title would be
unconditionally cleared. During dewikify, the link text would be
replaced with the page name, and in some circumstances the title would
also be set tor overwritten.

Now, the link text is only cleared by wikify when it equals the page
name, and set to the pagename by dewikify when it was empty, effectively
preserving any custom link text. For the title, this works similarly.

In addition, the wymeditor link editing code is changed to pull the page
name from the url instead of from the link text, as these are not
necessarily the same anymore. This also fixes link editing, which was
broken since urls in links were made relative.