Improve handling of private pages
The values for the private flag found in the database are inconsistent
because of different implementations, for ‘true’ three values can be
found; ‘true’, ‘1’ and ‘on’. With this implementation the value is
consistent; ‘on’ for true and ‘off’ for false.

Additionally, the check for a private page now happens in loadPage,
rather than in the page's show(). This helps to make sure that the flag
is always checked, and makes sure a 404 is returned.