wymeditor: Updated WYMeditor
Dialogs no longer appear in a popup, instead they are displayed in the page.
Fixing the problem that the dialogs of multiple editor instances always refer to the first editor instance.
Now a third party plugin is used for the image upload dialog instead of custom code.
Now a third party plugin is used for the site links dialog instead of custom code.