Do not overwrite .htaccess at install time
hypha.php would write out a .htaccess file at install time, but with
different contents from the one in git (that latter was modified, but
the version in hypha.php was not). This prevents writing the file at
install time, and makes sure that the file is included in a downloadable
hypha.php instead.

I suspect that the writing of the file was originally to allow
making local changes to the .htaccess file, without having those changes
reflected in a downloadable version, but that is a bit tricky to
implement without duplicating the contents, so for now .htaccess changes
just get included normally.