peer_reviewed_article: HTML-escape most values
In particular comments were interpreted as raw HTML, which poses an XSS
vulnerability (since comments can be posted by untrusted users), as well
as preventing non-malicious users from using < or > in their comments.

This comment escapes all comment content, as well as the author. The
article itself is obviously not escaped, since it is edited as HTML. The
sources and method fields are not escaped. These are not edited with an
HTML editor, but in practice they are used to write HTML into.

This fixes #105.