Fix libxml error logging
Since commit 1a5d24c (Hide unknown HTML tag errors from libxml), errors
from libxml are captured, some errors are supressed and the rest are
forwarded to PHP.

This used the error_log function, which causes errors to be directly to
the log, and passed an error level to it. However, error_log does not
actually accept an error level, but instead just logs strings as-is. The
error level passed was interpreted as a message type instead, and could
cause PHP to log the error without a newline, or send it through e-mail,
or generate errors because of unsupported values, all depending on the
error level passed.

This fixes the error logging by using the trigger_error function
instead. This function was probably intended initially and does accept
an error level parameter. It simply inserts the log message into the
normal error logging infrastructure, subject to the configured error
levels, error capturing, etc.

However, trigger_error only supports generating errors with E_USER_*.
Since E_USER_ERROR generates a fatal error that causes PHP to halt
execution (which *probably* is not what we want here), all levels are
changed to E_USER_WARNING as well.