Inprovements to the mailing module user interface
Sending out a mailing message can now only be done from a message's view. This avoids accidental sending of a message while composing it and encourages reviewing the message as it will be displayed before sending it out.

A test-send button was added that allows to send a mailing message to a custom email address with the purpose of reviewing the message as it arrives in ones inbox.

Finally, a warning message (Are you sure?) was added popping up before sending the bulk email.