In dewikify, support non-text content inside links
Before, dewikify looked only at the (absence of) text contents to decide
whether to insert the page title as the link text, and when replacing
the link with a span for inaccessible pages, it would only copy text
contents.

For links that did not contain any text, only other HTML (e.g. an
image), this would end up replacing the HTML with the pagename.