Let wymeditor set a base url for its iframe
This makes sure that any relative links in the edited HTML are resolved
in the same way as they would be in the page embedding the editor.

This modifies the wymeditor sources. The minified version is regenerated
from the regular javascript source using an online minifier:

  https://skalman.github.io/UglifyJS-online/

This is used with the default options, except that output.comments is
changed to false.