Fix dewikify error on private pages when not logged in
When a link to a private page is dewikified but the current user is not
logged in, the link is replaced by a non-clickable span with the same
text as the link.

This code had an error, where it called createElement on the page
element, rather than the document of the node to be replaced (and only
documents can create new nodes, not arbitrary elements). This caused a
fatal error, breaking the page in such cases.

This is fixed by this commit.