Dewikify all outgoing e-mails
This makes sure that any wikified links are dewikified, regardless of
how they end up in the mailing list.

This fixes at least links in digest e-mails and newsletters, possibly
others too.

This also removes an explicit dewikify from the newsletter confirmation
email, which is now no longer needed.

This is related to #179.