Simplify WymHTMLForm
Previously, HTMLForm contained a mapping of an HTML tagname to a field
type, which was then used by `updateFormField()` to decide how to handle
such a field. This was added so that WymHTMLForm could add support for
the `<editor>` tag.

With this commit, this mapping is removed, and WymHTMLForm just
overrides `updateFormField()` directly to add the required handling of
the `<editor>` tag. This simplifies things, and prepares for the next
commit.

This partially reverts commit 5269da3 (Extended HTMLForm with
WymHTMLForm in order to support the editor field).