Backport fix for php-dom-wrapper's html() method
In commit bb2afd6 (In dewikify, support non-text content inside links),
dewikify was changed to user `html()` instead of `text()` to include images and
other non-text elements as well in addition to text.

Now it turns out that `html()` does not actually include text, only non-text
child nodes, meaning dewikify only handles images and other non-text
contents and no longer handles text inside links.

This turns out to be a bug in php-dom-wrapper, see
https://github.com/scotteh/php-dom-wrapper/issues/15

This commit backports the upstream commit that fixes it, see
https://github.com/scotteh/php-dom-wrapper/commit/09000b4

Note that we cannot update to the latest version of php-dom-wrapper, since that
requires PHP 7.1 or higher

This fixes #186.