Wordwrap outgoing e-mail messages
RFC5322 specifies a max line length of 998 characters, which large bits
of wymeditor-generated HTML sometimes triggers. This does not always
pose problems, but some mailservers refuses such messages and some apply
hard wrapping (at exactly 998 characters), which can break the HTML
content.

To prevent that, we wrap the content ourselves, but only at whitespace,
which should be safe for HTML.

This wrapping might not handle more complex cases correctly (e.g.
multi-byte utf-8), so this solution might need to be improved later, but
for now it should at least be better than before.