In article, do not fire EVENT_PUBLIC_COMMENT for review comments
This seems like an oversight. Until recently, this event would not
actually do anyting (since review comments are always posted by users,
so never pending, so never trigger a confirmation e-mail), but since
commit 6bf3b52 (Implemented feedback to users on comments by either
visitors or users), the event would also send a notification message. It
seems this notification was intended for public comments only, not
review comments.

Not triggering the event for review comments prevents the notification
from being sent.