Adds option to preview a theme
A user needed to apply a theme in order to see or edit it, now a theme can be previewed and while previewing it, it can be edited as well.
A core-css file is loaded to ensure that the previewing can be canceled, just in case the preview would cause a viewing problem.