Make in-browser notifications consistent for article comments
Commit 9197655 (Improved feedback for visitor comments) attempted to
improve the feedback for public comments awaiting confirmation, but this
message would only be shown when the user clicked the confirm link twice
(so totally in the wrong place).

This commit makes sure to always either tell the user their comment is
posted (when they are logged in, or when they click the confirm link) or
should be confirmed (when they are not logged in).

This changes the string id "art-comment-received" to
"art-comment-pending" to be a bit more appropriate.