Added styling for texarea
The `normalize.css` forces font and font-size inheritance for all
elements, even for elements that normally define their own default font
(such as textarea). However, this means that a non-monospaced font is
used for textareas, which does not really look good.

This commit makes sure that a textarea gets a monospaced font again, at
the default font size, so a textarea looks like it would without any
font styling applied (at least in Firefox and Chrome). It would be even
nicer if we could say `font-family: initial` to revert to the default
font explicitly, but it seems that the "default textara font" is
actually defined in a browser-wide stylesheet of some sort, which is
also reverted when using `initial`. Hence this explicitly uses
`monospace`, which seems to work fine in practice.

This fixes #207