Let (de)wikify handle anchors in links
Before, it would treat the anchor part of a link as part of the
pagename and would create empty pages called e.g. `page_name#anchor`.
Since the links would still be dewikified back to their original value,
the anchors would actually work, but once these dummy pages are deleted,
dewikifying of the links breaks.

To simplify parsing the urls in wikify and dewikify, a regex is used
instead of adding another layer of explode calls.