HTMLForm: Store the form inside a dummy root element
The form elements are passed by the festival module in a dummy form tag
(which comes straight from the page XML). Previously, this dummy form
tag was stored as-is, and only the children were used. Now, a new dummy
root element is created (in its own document) and the children are
copied.

This prepares for extending the HTMLForm constructor to also accept the
form in different ways (e.g. an HTML string), but also removes the need
for the caller to copy the node when it should not be modified (and thus
prevents people forgetting to clone).