HTMLForm: Make Traversable
This lets HTMLForm implement the IteratorAggregate interface, which
makes it traversable. This means that you can iterate a HTMLForm using
e.g. foreach and get the individual (top-level) elements of the form.

This also means that you can pass an HTMLForm object to e.g.
dom-wrapper's append method and have it work as expected. This makes it
easier to work with forms and removes the need for callers to access
internal properties of the HTMLForm.