HTMLForm: Allow passing arbitrary contents to HTLMForm
Previously, the HTMLForm constructor would accept a DomElement, whose
children would be used as the form contents (and the wrapping DomElement
was ignored).

With this commit, whatever is passed to the constructor is
used as the form contents. This can be single DomElement, which will be
the single node in the form. It can also be a NodeList, array or any
Traversible containing DomElements, which will then all be used.
Finally, it can also be a html string, which will be parsed and used.

This relies on domwrapper's append function to handle all these cases,
which makes the code really compact.