Add command and argument fields on submit
Previously, these fields were added to the default form when it was
generated. This changes the hypha() javascript function to automatically
add these fields when submitting, rather than adding them in PHP. This
prepares for future changes where multiple forms can be used and removes
the need to add these fields in all those forms.

In addition, this restructures the hypha function a bit to prevent
adding these fields and setting values to them when the form is not
going to be submitted and to use jquery a bit more.