Store the default HTMLForm in HTMLDocument
This allows other code to retrieve the default form and modify it or
refer to it by id or otherwise. For this, the form must be generated at
the start of the request rather than in a post-processing function, so
this code is moved to index.php. To make sure the form can be modified
during the request, it is not added to the html until near the end of
the request (since adding it makes a copy).