Make the form to post explicit
Previously, the `hypha()` javascript function would always look up the
default form by name and use that for posting. With this commit the form
to use is always explicitly passed to `hypha()`.

Since there is still just a single default form, it is always explicitly
passed now. Most calls to `hypha()` go through the PHP `makeAction`
function, whose `$form` argument is optional and defaults to the default
form. Some `hypha()` calls are generated in javascript, which now use
`$(this).closest('form')` to get the form tag surrounding the submit
button (which is now the default form, but will automatically change to
a more specific form when it is added later).