WIP: No longer make the default form contain everything
This is a WIP, this commit must be updated once all datatypes have
switched to HTMLForm.

This means the default form is now an empty form, that can be used for
POST requests that just need to post a command (and optional argument).

Actual forms, with e.g. input elements, must now use their own form tag
to contain these input elements so they get POSTed.

To facilitate this, the HTMLForm class now automatically adds a form tag
if it is not already present and generates an id for it. Buttons that
submit forms have been updated to submit the right form rather than the
default form.

Javascript forms that do not use HTMLForm now add the form tag
themselves, as well as the CsrfToken field (which is only automatically
added to forms generated by PHP).