textpage: Invert if structure for form handling
Previously, this check for *no* errors and handle that inside the if,
and handle form errors after the if. However, the form errors case is
usually simpler (reshow the form and be done), so it makes more sense to
do that inside the if, putting the potentially more complicated valid
form handling below the if (where it is also less indented).