savePage: Properly detect page name conflicts
When renaming a page, `savePage()` would try to check whether the new
page name was already in use. However, this used a non-existing language
attribute, so it would never find any existing pages with the same name.
This fixes this by using the passed language instead.

This code seems to have been broken since the the first git revision.