peer_reviewed_article: Remove explicit approved status change
This status change was never actually used, since the statusMtx was
bypassed when displaying articles in REVIEW state (instead, an approve
button was shown for users that did not approve yet, and once enough
approves were collected, the status would be automatically changed).

However, this status change command was still available in the backend,
so a user could use a manually crafted HTTP request to change the status
from REVIEW to APPROVED, bypassing the normal checks for number of
approves and blocking comments.

This commit removes the handling for this request to prevent this. The
entry in statusMtx is kept, since the automatic status change does check
that.