Refactor newpage popup HTML generation
Previously, the HTML for this popup was composed in javascript, with
some parts being pregenerated in PHP. The code needed for this was not
very readable, and encoding was probably lacking in some areas.

With this commit, the full HTML for this popup is generated in PHP,
using the template and substitution method used for forms elsewhere.
Once this string is complete, the entire string is encoded using
json_encode, so it can be dynamically added to the DOM from javascript.