HyphaFile: Ensure always reading from the start
By default, `stream_get_contents` starts reading from the current
position. This passes extra arguments to make it rewind to the start of
the file first.

In practice, this would not be a problem, since files were read only
once after opening them, when the current position would still be at the
start (and writing already rewinds the pointer). However, this changes
makes it possible to read the file twice without problems as well.