HyphaFile: Make writing more error-proof
This changes the write method to write to a temporary file first, and
only when that file was completely succesfully written, the original
file is overwritten by the temporary file.

This approach ensures that when any problems occur while writing the new
file (such as a full disk, out of memory errors or other interruptions),
the write will be lost but the original file will be untouched and
intact, rather than truncated and corrupted.

The actual overwriting of the file uses rename(), which is an atomic
operation (at least on Linux, probably everywhere).

This does need additional changes to the lock method. Since the original
file is now overwritten, any other processes that are waiting for a
lockfile will get a lock on the *old* file (which is now overwritten so
no longer linked to a filename). This situation is detected by looking
at the number of filesystem links, and resolved by re-opening the file
and retrying the lock.

To preveng looping indefinitely, this is attempted a maximum of 10 times
(but note that this loops only once per released lock (since flock
blocks until the lock on the original file is obtained), it does not
spin 10 times quickly  and fail while just one other process is keeping
the lock for a bit longer).

Note that such errors currently result in an unhandled exception in most
cases, which is not handled so nicely yet (see #296), but at least there
is no data corruption.

This fixes #265.